+++ test/example/v3/callback/openapi.json
@@ -2,7 +2,7 @@
   "openapi": "3.0.0",
   "info": {
     "version": "1.0.0",
-    "title": "Callback Example"
+    "title": "Sample specification testing callback"
   },
   "paths": {
     "/streams": {
@@ -13,7 +13,8 @@
             "name": "callbackUrl",
             "in": "query",
             "required": true,
-            "description": "the location where data will be sent.  Must be network accessible\nby the source server\n",
+            "description":
+              "the location where data will be sent.  Must be network accessible\nby the source server\n",
             "schema": {
               "type": "string",
               "format": "uri",
@@ -28,12 +29,11 @@
               "application/json": {
                 "schema": {
                   "description": "subscription information",
-                  "required": [
-                    "subscriptionId"
-                  ],
+                  "required": ["subscriptionId"],
                   "properties": {
                     "subscriptionId": {
-                      "description": "this unique identifier allows management of the subscription",
+                      "description":
+                        "this unique identifier allows management of the subscription",
                       "type": "string",
                       "example": "2531329f-fb09-4ef7-887e-84e648214436"
                     }
@@ -67,10 +67,12 @@
                 },
                 "responses": {
                   "202": {
-                    "description": "Your server implementation should return this HTTP status code\nif the data was received successfully\n"
+                    "description":
+                      "Your server implementation should return this HTTP status code\nif the data was received successfully\n"
                   },
                   "204": {
-                    "description": "Your server should return this HTTP status code if no longer interested\nin further updates\n"
+                    "description":
+                      "Your server should return this HTTP status code if no longer interested\nin further updates\n"
                   }
                 }
               }
